Extract persona/review lifecycle tools from MCP server.ts into persona-tools.ts
Move 5 review lifecycle tool registrations (review_status, dispatch_complete_review,
get_parent_context, dispatch_get_recheck_context, dispatch_cancel_recheck) into a
dedicated persona-tools.ts module. Reduces server.ts from 1584 to 1377 lines.
